+++ crates/rslint_core/CHANGELOG.md
@@ -5,10 +5,10 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
-- Added "constructor-super" rule
 
 ### Added
 
+- Added "constructor-super" rule
 - Added `util::StyleExt` for stylistic linting
 - Added `Fixer::delete_multiple`
 - Added more functions to `util::StyleExt`